\name{addbrush}
\alias{addbrush}
\title{Add brush to plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This adds a rectangle to a ggplot plot indicating the brush position
}
\usage{addbrush(plot, x,y, width=0.5, height=0.5, just=c("left", "top"), fill=NA, col="black")}
\arguments{
\item{plot}{plot object}
\item{x}{x position of brush}
\item{y}{y position of brush}
\item{width}{width of brush}
\item{height}{height of brush}
\item{just}{which corner of brush should be determined by x and y position}
\item{fill}{fill colour for brush (use \code{\link[ggplot]{alpha}} for alpha blending)}
\item{col}{outline colour of brush}
}

\details{}

\examples{}
\keyword{hplot}
