% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addbrush}
\alias{addbrush}
\title{Add brush to plot
This adds a rectangle to a ggplot plot indicating the brush position.}
\usage{
addbrush(
  plot,
  x,
  y,
  width = 0.5,
  height = 0.5,
  just = c("left", "top"),
  fill = "transparent",
  col = "black"
)
}
\arguments{
\item{plot}{plot object}

\item{x}{x position of brush}

\item{y}{y position of brush}

\item{width}{width of brush}

\item{height}{height of brush}

\item{just}{which corner of brush should be determined by x and y position}

\item{fill}{fill colour for brush.  Use ggplot-alpha for alpha blending.}

\item{col}{outline colour of brush}
}
\description{
Add brush to plot
This adds a rectangle to a ggplot plot indicating the brush position.
}
\author{
Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{hplot}
