\name{axesViewport}
\alias{axesViewport}
\title{Axes viewport}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Construct viewport for axes.
}
\usage{axesViewport(axes, axislocation)}
\arguments{
\item{axes}{describe display object}
\item{axislocation}{plot}
}

\details{}

\examples{}
\keyword{internal}
