% Generated by roxygen2 (4.0.2): do not edit by hand
\name{axesViewport}
\alias{axesViewport}
\title{Axes viewport
Construct viewport for axes.}
\usage{
axesViewport(axes, axislocation)
}
\arguments{
\item{axes}{describe display object}

\item{axislocation}{location of axes (as x and y position in npc coordinates, ie. between 0 and 1)}
}
\description{
Axes viewport
Construct viewport for axes.
}
\author{
Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}

