\name{compact_pcp}
\alias{compact_pcp}
\title{Compact pcp data}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
A parallel coordinates is written out as a series of 1D dotplots.  This function
}
\usage{compact_pcp(data)}
\arguments{
\item{data}{data}
}

\details{}

\examples{}
\keyword{internal}
