\name{compact_pcp}
\alias{compact_pcp}
\title{Compact pcp data...}
\usage{compact_pcp(data)}
\description{Compact pcp data
A parallel coordinates is written out as a series of 1D dotplots.  This function
compacts it back into one dataset.}
\author{Hadley Wickham h.wickham [at] gmail.com}
\keyword{internal}
\arguments{\item{data}{}}
