\name{compact_scatmat}
\alias{compact_scatmat}
\title{Compact scatmat data...}
\usage{compact_scatmat(data)}
\description{Compact scatmat data
A scatter plot matrix is made from the information from the 1Dplots. This function collects only that data.}
\author{Barret Schloerke schloerke [at] gmail.com}
\keyword{internal}
\arguments{\item{data}{}}
