\name{compact_timeseries}
\alias{compact_timeseries}
\title{Compact time series data
A time series plot is made from scatterplots with a common x axis.
This function pulls the correct information out of the data.}
\usage{
  compact_timeseries(data)
}
\arguments{
  \item{data}{}
}
\description{
  Compact time series data A time series plot is made from
  scatterplots with a common x axis. This function pulls
  the correct information out of the data.
}
\author{
  Barret Schloerke \email{bigbear@iastate.edu}
}
\keyword{internal}

