\name{dd_clean_plot}
\alias{dd_clean_plot}
\title{Clean plot data structure}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Cleans up plot data structure into consistent, easy to use data structure
}
\usage{dd_clean_plot(dd, n=1)}
\arguments{
\item{dd}{dd object}
\item{n}{plot number}
}

\details{}

\examples{}
\keyword{internal}
