\name{dd_defaults}
\alias{dd_defaults}
\title{Describe display plot defaults}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Gather overall plot defaults for specified plot
}
\usage{dd_defaults(dd, n=1)}
\arguments{
\item{dd}{list of values from describe display}
\item{n}{plot number, defaults to first plot}
}

\details{}

\examples{}
\keyword{internal}
