\name{dd_plot_class}
\alias{dd_plot_class}
\title{Describe display plot class}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Compute valid R class name for given plot type
}
\usage{dd_plot_class(projection)}
\arguments{
\item{projection}{list of values from describe display}
\item{}{plot number, defaults to first plot}
}

\details{}

\examples{}
\keyword{internal}
