\name{dd_points}
\alias{dd_points}
\title{Describe display points data}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Retrieves the describe display points data for the given plot number.
}
\usage{dd_points(dd, n=1)}
\arguments{
\item{dd}{list of values from describe display}
\item{n}{plot number, defaults to first plot}
}
\value{data frame suitable for plotting}
\details{}

\examples{}
\keyword{internal}
