\name{dd_tour_axes}
\alias{dd_tour_axes}
\title{Describe display tour axis}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Return representation of axes for specified plot
}
\usage{dd_tour_axes(plot)}
\arguments{
\item{plot}{list of values from describe display}
\item{}{plot number, defaults to first plot}
}

\details{}

\examples{}
\keyword{internal}
