% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dd_clean_plot}
\alias{dd_clean_plot}
\title{Clean plot data structure
Cleans up plot data structure into consistent, easy to use data structure}
\usage{
dd_clean_plot(dd, n = 1)
}
\arguments{
\item{dd}{dd object}

\item{n}{plot number}
}
\description{
Clean plot data structure
Cleans up plot data structure into consistent, easy to use data structure
}
\author{
Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}
