\name{dd_defaults}
\alias{dd_defaults}
\title{Describe display plot defaults...}
\usage{dd_defaults(dd, n=1)}
\description{Describe display plot defaults
Gather overall plot defaults for specified plot}
\author{Hadley Wickham \email{h.wickham@gmail.com}}
\keyword{internal}
\arguments{\item{dd}{list of values from describe display}
\item{n}{plot number, defaults to first plot}}
