% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dd_edges}
\alias{dd_edges}
\title{Describe display edge data
Retrieves the describe display edge data for the given plot number.}
\usage{
dd_edges(dd, n = 1)
}
\arguments{
\item{dd}{list of values from describe display}

\item{n}{plot number, defaults to first plot}
}
\value{
data frame suitable for plotting
}
\description{
Describe display edge data
Retrieves the describe display edge data for the given plot number.
}
\author{
Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}
