\name{dd_example}
\alias{dd_example}
\title{Load example describe display file...}
\usage{dd_example(name)}
\description{Load example describe display file
Load example describe display file included with package.}
\details{These are mainly used for testing.}
\author{Hadley Wickham \email{h.wickham@gmail.com}}
\keyword{internal}
\arguments{\item{name}{name of example}}
\examples{a <- dd_example("xyplot")}
