\name{dd_plot_class}
\alias{dd_plot_class}
\title{Describe display plot class...}
\usage{dd_plot_class(projection)}
\description{Describe display plot class
Compute valid R class name for given plot type}
\author{Hadley Wickham \email{h.wickham@gmail.com}}
\keyword{internal}
\arguments{\item{projection}{type of projection that should be used}}
