\name{dd_points}
\alias{dd_points}
\title{Describe display points data...}
\usage{dd_points(dd, n=1)}
\description{Describe display points data
Retrieves the describe display points data for the given plot number.}
\value{data frame suitable for plotting}
\author{Hadley Wickham \email{h.wickham@gmail.com}}
\keyword{internal}
\arguments{\item{dd}{list of values from describe display}
\item{n}{plot number, defaults to first plot}}
