\name{dd_tour_axes}
\alias{dd_tour_axes}
\title{Describe display tour axis...}
\usage{dd_tour_axes(plot)}
\description{Describe display tour axis
Return representation of axes for specified plot}
\author{Hadley Wickham h.wickham [at] gmail.com}
\keyword{internal}
\arguments{\item{plot}{list of information of a plot}}
