\name{ddpanelGrob}
\alias{ddpanelGrob}
\title{Panel grob...}
\usage{ddpanelGrob(panel, axislocation=c(0.1, 0.1),
    axis.gp=gpar(col = "black"),
    background.color=grey90)}
\description{Panel grob
Construct grob for single panel.}
\author{Hadley Wickham \email{h.wickham@gmail.com}}
\keyword{internal}
\arguments{\item{panel}{describe display object}
\item{axis}{location, x and y position}
\item{axisgp}{frame grob containing all panels, note that this does not contain the title or border}
\item{background.color}{color of in the background of the plot}}
