\name{DescribeDisplay-package}
\alias{DescribeDisplay-package}
\alias{DescribeDisplay}
\docType{package}
\title{
R interface to DescribeDisplay (GGobi plugin)
}
\description{
Produce publication quality graphics from output of GGobi's describe display plugin
}
\details{
\tabular{ll}{
Package: \tab DescribeDisplay\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2009-03-31\cr
License: \tab BSD\cr
LazyData: \tab yes\cr
}
}
\author{
Hadley Wickham <hadley@iastate.edu>, Di Cook <dicook@iastate.edu>, Andreas Buja <buja@wharton.upenn.edu>, Barret Schloerke <bigbear@iastate.edu>
Maintainer: Hadley Wickham <h.wickham@gmail.com>
}
\references{
http://www.ggobi.org/describe-display/
http://www.had.co.nz/ggplot2/
}
\keyword{ package }
\examples{
	example(ggplot.dd)
	example(ggplot.scatmat)
	example(ggplot.parcoords)
	example(ggplot.timeseries)
}
