\name{geom_axis}
\alias{geom_axis}
\alias{GeomAxis}
\title{Geom Axis
A special ggplot2 geom for drawing the tour axes}
\usage{
  geom_axis(...)
}
\arguments{
  \item{...}{should include data, location, aes\_string
  information}
}
\description{
  Geom Axis A special ggplot2 geom for drawing the tour
  axes
}
\author{
  Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}

