\name{ggplot.parcoords}
\alias{ggplot.parcoords}
\title{Create a nice plot for parallel coordinates plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create a nice looking plot complete with axes using ggplot.
}
\usage{ggplot.parcoords(data, ...)}
\arguments{
\item{data}{plot to display}
\item{...}{other (currently) unused arguments}
}

\details{}

\examples{ggplot(dd_example("pcp"))
ggplot(dd_example("pcp"), size = 1)}
\keyword{hplot}
