\name{ggplot.timeseries}
\alias{ggplot.timeseries}
\title{Create nice plots for a time series...}
\usage{ggplot.timeseries(data, edges=FALSE, ...)}
\description{Create nice plots for a time series
Create nice looking plots complete with axes using ggplot.  Produces graphics with a uniform x axis.}
\author{Barret Schloerke schloerke [at] gmail.com}
\keyword{hplot}
\arguments{\item{plot}{to display}
\item{edges}{Boolean operator to tell whether to try to force the edges or not.  Will not work to remove the edges.}
\item{other}{(currently) unused arguments}}
