\name{nulldefault}
\alias{nulldefault}
\title{Null default}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convienece function for setting defaults for null values
}
\usage{nulldefault(x, default)}
\arguments{
\item{x}{value}
\item{default}{default to use if value is null}
}

\details{}

\examples{}
\keyword{internal}
