\name{panelGrob}
\alias{panelGrob}
\title{Panel grob}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Construct grob for single panel.
}
\usage{panelGrob(panel,axislocation = c(0.1, 0.1), axis.gp = gpar(col="black"), background.color="grey90")}
\arguments{
\item{panel}{describe display object}
\item{axislocation}{plot}
\item{axis.gp}{axis location, x and y position}
\item{background.color}{}
}

\details{}

\examples{}
\keyword{internal}
