\name{plot.dd_plot}
\alias{plot.dd_plot}
\title{Plot a dd plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convenient method to draw a single panel.
}
\usage{plot.dd_plot(x, ..., axislocation = c(0.1, 0.1), axisgp=gpar(col="black"), background.color = "grey90")}
\arguments{
\item{x}{object to plot}
\item{...}{axis location, x and y position}
\item{axislocation}{}
\item{axisgp}{}
\item{background.color}{}
}

\details{This is mainly used for bug testing so that you can pull out a single
panel quickly and easily.}

\examples{}
\keyword{hplot}
