% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ddplot}
\alias{plot.ddplot}
\title{Plot a dd plot
Convenient method to draw a single panel.}
\usage{
\method{plot}{ddplot}(x, ..., axislocation = c(0.1, 0.1), axis.gp = gpar(col
  = "black"), background.color = "grey90")
}
\arguments{
\item{x}{object to plot}

\item{...}{(not used)}

\item{axislocation}{location of axes (as x and y position in npc coordinates, ie. between 0 and 1)}

\item{axis.gp}{frame grob containing all panels, note that this does not contain the title or border}

\item{background.color}{color of in the background of the plot}
}
\description{
This is mainly used for bug testing so that you can pull out a single
panel quickly and easily.
}
\examples{
scatmat <- dd_example("scatmat")
plot(scatmat)
plot(scatmat$plots[[1]])
plot(scatmat$plots[[3]])
plot(scatmat$plots[[4]])
}
\author{
Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{hplot}
