\name{print.dd}
\alias{print.dd}
\title{Print dd object
Use str to print out human readable describe display object}
\usage{
  print.dd(x, ...)
}
\arguments{
  \item{x}{dd object}

  \item{...}{not used}
}
\description{
  Print dd object Use str to print out human readable
  describe display object
}
\author{
  Hadley Wickham h.wickham [at] gmail.com
}
\keyword{internal}

