\name{range01}
\alias{range01}
\title{Scale the values by range
Divide the values of the objects by the range of values}
\usage{
range01(x)
}
\arguments{
  \item{x}{values to be worked on}
}
\description{
  Scale the values by range Divide the values of the
  objects by the range of values
}
\author{
  Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}
