% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zeeThemAll}
\alias{zeeThemAll}
\title{Run All Examples
Will run all examples within the package}
\usage{
zeeThemAll()
}
\description{
Run All Examples
Will run all examples within the package
}
\author{
Barret Schloerke bigbear@iastate.edu
}
\keyword{hplot}

