% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis-grob.R
\name{axesGrob}
\alias{axesGrob}
\title{Axes grob
Construct grob for axes.}
\usage{
axesGrob(axes, gp = gpar(col = "black"))
}
\arguments{
\item{axes}{describe display object}

\item{gp}{arguments to the axes}
}
\description{
Axes grob
Construct grob for axes.
}
\author{
Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}
