\name{axesGrob}
\alias{axesGrob}
\title{Axes grob...}
\usage{axesGrob(axes, gp=gpar(col = "black"))}
\description{Axes grob
Construct grob for axes.}
\author{Hadley Wickham \email{h.wickham@gmail.com}}
\keyword{internal}
\arguments{\item{axes}{describe display object}
\item{gp}{arguments to the axes}}
