\name{axesGrob}
\alias{axesGrob}
\title{Axes grob
Construct grob for axes.}
\usage{
axesGrob(axes, gp = gpar(col = "black"))
}
\arguments{
  \item{axes}{describe display object}

  \item{gp}{arguments to the axes}
}
\description{
  Axes grob Construct grob for axes.
}
\author{
  Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}
