\name{axesViewport}
\alias{axesViewport}
\title{Axes viewport...}
\usage{axesViewport(axes, axislocation)}
\description{Axes viewport
Construct viewport for axes.}
\author{Hadley Wickham \email{h.wickham@gmail.com}}
\keyword{internal}
\arguments{\item{axes}{describe display object}
\item{axislocation}{location of axes (as x and y position in npc coordinates, ie. between 0 and 1)}}
