\name{compact_pcp}
\alias{compact_pcp}
\title{Compact pcp data
A parallel coordinates is written out as a series of 1D dotplots.  This function
compacts it back into one dataset.}
\usage{
  compact_pcp(data)
}
\arguments{
  \item{data}{data to pull points from}
}
\description{
  Compact pcp data A parallel coordinates is written out as
  a series of 1D dotplots.  This function compacts it back
  into one dataset.
}
\author{
  Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}

