\name{compact_scatmat}
\alias{compact_scatmat}
\title{Compact scatmat data
A scatter plot matrix is made from the information from the 1Dplots. This function collects only that data.}
\usage{
compact_scatmat(data)
}
\arguments{
  \item{data}{}
}
\description{
  Compact scatmat data A scatter plot matrix is made from
  the information from the 1Dplots. This function collects
  only that data.
}
\author{
  Barret Schloerke \email{bigbear@iastate.edu}
}
\keyword{internal}
