% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scatmat.R
\name{compact_scatmat}
\alias{compact_scatmat}
\title{Compact scatmat data
A scatter plot matrix is made from the information from the 1Dplots. This function collects only that data.}
\usage{
compact_scatmat(data)
}
\arguments{
\item{data}{data to display}
}
\description{
Compact scatmat data
A scatter plot matrix is made from the information from the 1Dplots. This function collects only that data.
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{internal}
