% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-timeseries.R
\name{compact_timeseries}
\alias{compact_timeseries}
\title{Compact time series data
A time series plot is made from scatterplots with a common x axis.
This function pulls the correct information out of the data.}
\usage{
compact_timeseries(data)
}
\arguments{
\item{data}{data to display}
}
\description{
Compact time series data
A time series plot is made from scatterplots with a common x axis.
This function pulls the correct information out of the data.
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{internal}

