% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dd_defaults}
\alias{dd_defaults}
\title{Describe display plot defaults
Gather overall plot defaults for specified plot}
\usage{
dd_defaults(dd, n = 1)
}
\arguments{
\item{dd}{list of values from describe display}

\item{n}{plot number, defaults to first plot}
}
\description{
Describe display plot defaults
Gather overall plot defaults for specified plot
}
\author{
Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}
