\name{dd_edges}
\alias{dd_edges}
\title{Describe display edge data...}
\usage{dd_edges(dd, n=1)}
\description{Describe display edge data
Retrieves the describe display edge data for the given plot number.}
\value{data frame suitable for plotting}
\author{Hadley Wickham \email{h.wickham@gmail.com}}
\keyword{internal}
\arguments{\item{dd}{list of values from describe display}
\item{n}{plot number, defaults to first plot}}
