\name{dd_example}
\alias{dd_example}
\title{Load example describe display file
Load example describe display file included with package.}
\usage{
dd_example(name)
}
\arguments{
  \item{name}{name of example}
}
\description{
  These are mainly used for testing.
}
\examples{
a <- dd_example("xyplot")
}
\author{
  Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}
