% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dd_plot_class}
\alias{dd_plot_class}
\title{Describe display plot class
Compute valid R class name for given plot type}
\usage{
dd_plot_class(projection)
}
\arguments{
\item{projection}{type of projection that should be used}
}
\description{
Describe display plot class
Compute valid R class name for given plot type
}
\author{
Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}
