% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dd_tour_axes}
\alias{dd_tour_axes}
\title{Describe display tour axis
Return representation of axes for specified plot}
\usage{
dd_tour_axes(plot)
}
\arguments{
\item{plot}{list of information of a plot}
}
\description{
Describe display tour axis
Return representation of axes for specified plot
}
\author{
Hadley Wickham h.wickham [at] gmail.com
}
\keyword{internal}
