% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ddpanelGrob}
\alias{ddpanelGrob}
\title{Panel grob
Construct grob for single panel.}
\usage{
ddpanelGrob(
  panel,
  axislocation = c(0.1, 0.1),
  axis.gp = gpar(col = "black"),
  background.color = "grey90"
)
}
\arguments{
\item{panel}{describe display object}

\item{axis.gp}{frame grob containing all panels, note that this does not contain the title or border}

\item{background.color}{color of in the background of the plot}

\item{axis}{location, x and y position}
}
\description{
Panel grob
Construct grob for single panel.
}
\author{
Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}
