\docType{package}
\name{describedisplayreshape}
\alias{describedisplayreshape}
\alias{describedisplayreshape-package}
\title{R interface to DescribeDisplay (GGobi plugin).}
\description{
  Produce publication quality graphics from output of
  GGobi's describe display plugin
}
\references{
  \url{http://www.ggobi.org/describe-display}
  \url{http://www.ggobi.org/}
  \url{http://www.had.co.nz/ggplot2/}
}
