% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescribeDisplay-package.R
\docType{package}
\name{describedisplayreshape}
\alias{describedisplayreshape}
\title{R interface to DescribeDisplay (GGobi plugin).}
\description{
Produce publication quality graphics from output of GGobi's describe
display plugin
}
\references{
\url{http://ggobi.org}
  \url{https://ggplot2.tidyverse.org}
}
