\name{geom_axis}
\alias{geom_axis}
\title{Geom Axis.}
\usage{
geom_axis(axis, location, ...)
}
\arguments{
  \item{...}{should include data, location, aes\_string
  information}
}
\description{
  A special ggplot2 geom for drawing the tour axes
}
\examples{
library(ggplot2)
print(ggplot(dd_example("tour2d")))
}
\author{
  Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}
