\name{geom_axis}
\title{Geom Axis...}
\usage{geom_axis(...)}
\description{Geom Axis
A special ggplot2 geom for drawing the tour axes}
\author{Hadley Wickham \email{h.wickham@gmail.com}}
\keyword{internal}
\alias{GeomAxis}
\arguments{\item{...}{should include data, location, aes\_string information}}
