\name{GeomAxis}
\title{GeomAxis...}
\description{GeomAxis
A special ggplot2 geom for drawing the tour axes}
\keyword{internal}
\author{Hadley Wickham \email{h.wickham@gmail.com}}
\alias{geom_axis}
\arguments{\item{Geom}{previous information}
\item{new}{function to set up the axis}}
\examples{print(ggplot(dd_example("tour2d")))}
