\name{ggplot.dd}
\alias{ggplot.dd}
\title{Create a nice plot...}
\usage{ggplot.dd(data, ...)}
\description{Create a nice plot
Create a nice looking plot complete with axes using ggplot.}
\author{Hadley Wickham h.wickham [at] gmail.com}
\keyword{hplot}
\arguments{\item{data}{plot to display, object created by \code{dd_load()}}
\item{other}{not used}}
\examples{ggplot(dd_example("xyplot"))}
