\name{ggplot.dd}
\alias{ggplot.dd}
\title{Create a nice plot
Create a nice looking plot complete with axes using ggplot.}
\usage{
  ggplot.dd(data, ...)
}
\arguments{
  \item{data}{plot to display, object created by
  \code{dd_load()}}

  \item{...}{not used}
}
\description{
  Create a nice plot Create a nice looking plot complete
  with axes using ggplot.
}
\examples{
library(ggplot2)
print(example(ggplot.ddplot))
print(example(ggplot.histogram))
print(example(ggplot.barplot))
}
\author{
  Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{hplot}

