\name{ggplot.dd}
\alias{ggplot.dd}
\title{Create a nice plot...}
\usage{ggplot.dd(data, ...)}
\description{Create a nice plot
Create a nice looking plot complete with axes using ggplot.}
\author{Hadley Wickham \email{h.wickham@gmail.com}}
\keyword{hplot}
\arguments{\item{data}{plot to display, object created by \code{dd_load()}}
\item{...}{not used}}
\examples{print(example(ggplot.ddplot))
print(example(ggplot.histogram))
print(example(ggplot.barplot))}
