\name{ggplot.parcoords}
\alias{ggplot.parcoords}
\title{Create a nice plot for parallel coordinates plot...}
\usage{ggplot.parcoords(data, ...)}
\description{Create a nice plot for parallel coordinates plot
Create a nice looking plot complete with axes using ggplot.}
\author{Barret Schloerke schloerke [at] gmail.com}
\keyword{hplot}
\arguments{\item{plot}{to display}
\item{other}{(currently) unused arguments}}
\examples{ggplot(dd_example("pcp"))
ggplot(dd_example("pcp"), size = 1)}
