% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scatmat.R
\name{ggplot.scatmat}
\alias{ggplot.scatmat}
\title{Create a nice plots in a scatter plot matrix
Create a nice looking plots in a matrix.  The 1d sections along
the diagonal have a smooth density while the values are compared
to eachother within the matrix.}
\usage{
\method{ggplot}{scatmat}(data, ...)
}
\arguments{
\item{data}{data to display}

\item{...}{(currently) unused arguments}
}
\description{
Create a nice plots in a scatter plot matrix
Create a nice looking plots in a matrix.  The 1d sections along
the diagonal have a smooth density while the values are compared
to eachother within the matrix.
}
\examples{
library(ggplot2)
print(ggplot(dd_example("scatmat")))
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}
