% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-timeseries.R
\name{ggplot.timeseries}
\alias{ggplot.timeseries}
\title{Create nice plots for a time series
Create nice looking plots complete with axes using ggplot.  Produces graphics with a uniform x axis.}
\usage{
\method{ggplot}{timeseries}(data, edges = FALSE, ...)
}
\arguments{
\item{data}{to display}

\item{edges}{Boolean operator to tell whether to try to force the edges or not.  Will not work to remove the edges.}

\item{...}{(currently) unused arguments}
}
\description{
Create nice plots for a time series
Create nice looking plots complete with axes using ggplot.  Produces graphics with a uniform x axis.
}
\examples{
library(ggplot2)
print(ggplot(dd_example("timeseries")))
print(ggplot(dd_example("timeseries"), edges = TRUE))
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}
