\name{panelGrob}
\alias{panelGrob}
\title{Panel grob}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Construct grob for single panel.
}
\usage{panelGrob(panel,axislocation = c(0.1, 0.1))}
\arguments{
\item{panel}{describe display object}
\item{axislocation}{plot}
\item{}{axis location, x and y position}
}

\details{}

\examples{}
\keyword{internal}
