\name{removehiddens}
\alias{removehiddens}
\title{Remove hidden points
Will remove all hidden points from the plot.}
\usage{
  removehiddens(d)
}
\arguments{
  \item{d}{ddplot object}
}
\description{
  Remove hidden points Will remove all hidden points from
  the plot.
}
\author{
  Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{manip}

