% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zeeThemAll}
\alias{zeeThemAll}
\title{Run All Examples
Will run all examples within the package}
\usage{
zeeThemAll()
}
\description{
Run All Examples
Will run all examples within the package
}
\author{
Barret Schloerke schloerke@gmail.com
}
\keyword{hplot}

